Added more composition methods
